requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
delayed-assert==0.3.6
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-decouple==3.5
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
